requirements.txt
web3==6.11.3
pandas==2.1.4
psycopg2-binary==2.9.9
pymongo==4.6.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
streamlit==1.28.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0 